+++ monitor/monitor_ping.go
@@ -93,8 +93,13 @@ func (p Ping) MarshalJSON() ([]byte, error) {
 type PingDetails struct {
 	Hostname   string `json:"hostname"`
 	PacketSize int    `json:"packetSize"`
-	// Timeout is an optional request timeout in seconds.
-	Timeout *int64 `json:"timeout"`
+	// Timeout is an optional request timeout in seconds. Although the server
+	// stores it in a floating point column, it rounds the value to whole
+	// seconds before storing it for ping monitors and rejects anything
+	// outside 1 to 300 seconds or below the per-request ping timeout, which
+	// defaults to 2. A fractional value therefore does not round-trip
+	// unchanged.
+	Timeout *float64 `json:"timeout"`
 	// DomainExpiryNotification enables domain expiry notifications
 	// for the monitored domain.
 	DomainExpiryNotification bool `json:"domainExpiryNotification"`
