+++ monitor/monitor_real_browser.go
@@ -95,8 +95,12 @@ func (r RealBrowser) MarshalJSON() ([]byte, error) {
 
 // RealBrowserDetails contains realbrowser-specific monitor configuration.
 type RealBrowserDetails struct {
-	URL                      string   `json:"url"`
-	Timeout                  int64    `json:"timeout"`
+	URL string `json:"url"`
+	// Timeout is the request timeout in seconds. The server stores it in a
+	// floating point column, so fractional values round-trip unchanged, but
+	// the realbrowser check never reads the column: it derives its timeout
+	// from 80 percent of the interval instead.
+	Timeout                  float64  `json:"timeout"`
 	IgnoreTLS                bool     `json:"ignoreTls"`
 	MaxRedirects             int      `json:"maxredirects"`
 	AcceptedStatusCodes      []string `json:"accepted_statuscodes"`
